Clean up test suite names
